pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "Template for mcp server with connection to database"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi==0.116.1",
    "mcp==1.8.0",
    "pandas==2.3.1",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "uvicorn==0.35.0",
    "asyncpg>=0.30.0",
    "faiss-cpu>=1.11.0.post1",
    "rapidfuzz>=3.13.0",
    "numpy==1.26.4",
    "nltk>=3.8.1,<4.0",
    "inflect>=7.0.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]